ddd refator
