+++ src/main/java/ar/com/ddd/ddd_architecture/catalog/application/DeleteBookByIdUseCase.java
@@ -1,0 +1,18 @@
+package ar.com.ddd.ddd_architecture.catalog.application;
+
+import ar.com.ddd.ddd_architecture.catalog.domain.providers.BookDAOProvider;
+import ar.com.ddd.ddd_architecture.shared.utils.UseCase;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+
+@UseCase
+@Slf4j
+@RequiredArgsConstructor
+public class DeleteBookByIdUseCase {
+    private final BookDAOProvider bookDAO;
+
+    public void execute(Long id){
+        this.bookDAO.delete(id);
+    }
+}
