+++ src/main/java/ar/com/ddd/ddd_architecture/catalog/domain/providers/BookClientProvider.java
@@ -1,0 +1,7 @@
+package ar.com.ddd.ddd_architecture.catalog.domain.providers;
+
+import ar.com.ddd.ddd_architecture.catalog.domain.models.BookInformationModel;
+
+public interface BookClientProvider {
+    public BookInformationModel search(String isbn);
+}
