--- src/main/java/ar/com/ddd/ddd_architecture/catalog/domain/usecases/DeleteBookByIdUseCase.java
@@ -1,18 +1,0 @@
-package ar.com.ddd.ddd_architecture.catalog.domain.usecases;
-
-import ar.com.ddd.ddd_architecture.catalog.domain.repository.BookDAO;
-import ar.com.ddd.ddd_architecture.common.utils.UseCase;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-
-@UseCase
-@Slf4j
-@RequiredArgsConstructor
-public class DeleteBookByIdUseCase {
-    private final BookDAO bookDAO;
-
-    public void execute(Long id){
-        this.bookDAO.delete(id);
-    }
-}
