+++ src/main/java/ar/com/ddd/ddd_architecture/catalog/infrastructure/client/OpenClientResultsJson.java
@@ -1,0 +1,9 @@
+package ar.com.ddd.ddd_architecture.catalog.infrastructure.client;
+
+import java.util.List;
+
+public record OpenClientResultsJson(List<String> publishers,
+                                    String title,
+                                    List<String> isbn_13,
+                                    int revisions) {
+}
