+++ src/main/java/ar/com/ddd/ddd_architecture/catalog/infrastructure/entity/BookEntity.java
@@ -1,0 +1,25 @@
+package ar.com.ddd.ddd_architecture.catalog.infrastructure.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Table(name = "books")
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class BookEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String isbn;
+    private String name;
+}
