--- src/main/java/ar/com/ddd/ddd_architecture/catalog/infrastructure/persistence/BookPersistenceDAO.java
@@ -1,8 +1,0 @@
-package ar.com.ddd.ddd_architecture.catalog.infrastructure.persistence;
-
-
-import ar.com.ddd.ddd_architecture.catalog.domain.repository.BookDAO;
-
-public interface BookPersistenceDAO extends BookDAO{
-    
-}
