+++ src/main/java/ar/com/ddd/ddd_architecture/catalog/infrastructure/persistence/repository/BookHibernateRepository.java
@@ -2,8 +2,8 @@ package ar.com.ddd.ddd_architecture.catalog.infrastructure.persistence.repositor
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import ar.com.ddd.ddd_architecture.catalog.infrastructure.entity.Book;
+import ar.com.ddd.ddd_architecture.catalog.infrastructure.entity.BookEntity;
 
-public interface BookHibernateRepository extends JpaRepository<Book, Long> {
+public interface BookHibernateRepository extends JpaRepository<BookEntity, Long> {
 
 }
