--- src/main/java/ar/com/ddd/ddd_architecture/common/dto/GenericResponse.java
@@ -1,19 +1,0 @@
-package ar.com.ddd.ddd_architecture.common.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class GenericResponse<T> {
-    private Integer code;
-    private String message;
-    private T data;
-}
