--- src/main/java/ar/com/ddd/ddd_architecture/common/utils/DataAccessObject.java
@@ -1,15 +1,0 @@
-package ar.com.ddd.ddd_architecture.common.utils;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import org.springframework.stereotype.Component;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-@Component
-public @interface DataAccessObject {
-    
-}
