--- src/main/java/ar/com/ddd/ddd_architecture/common/utils/Provider.java
@@ -1,14 +1,0 @@
-package ar.com.ddd.ddd_architecture.common.utils;
-import org.springframework.stereotype.Component;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-@Component
-public @interface Provider {
-    
-}
