--- src/main/java/ar/com/ddd/ddd_architecture/common/utils/UseCase.java
@@ -1,13 +1,0 @@
-package ar.com.ddd.ddd_architecture.common.utils;
-
-import org.springframework.stereotype.Service;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-@Service
-public @interface UseCase {
-}
