--- src/main/java/ar/com/ddd/ddd_architecture/exceptions/GenericException.java
@@ -1,19 +1,0 @@
-package ar.com.ddd.ddd_architecture.exceptions;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-@Data
-@EqualsAndHashCode(callSuper = false)
-public class GenericException extends RuntimeException {
-
-    private Integer code;
-    private String detail;
-    private String message;
-
-    public GenericException(Integer code, String message, String detail) {
-        super(message);
-        this.message = message;
-        this.code = code;
-        this.detail = detail;
-    }
-}
