--- src/main/java/ar/com/ddd/ddd_architecture/exceptions/InternalException.java
@@ -1,11 +1,0 @@
-package ar.com.ddd.ddd_architecture.exceptions;
-
-import org.springframework.http.HttpStatus;
-
-public class InternalException extends GenericException {
-    
-    private static final String MESSAGE = "INTERNAL_SERVER_ERROR";
-    public InternalException(String detail) {
-        super(HttpStatus.INTERNAL_SERVER_ERROR.value(), MESSAGE, detail);
-    }
-}
