--- src/main/java/ar/com/ddd/ddd_architecture/exceptions/NotFoundException.java
@@ -1,11 +1,0 @@
-package ar.com.ddd.ddd_architecture.exceptions;
-
-import org.springframework.http.HttpStatus;
-
-public class NotFoundException extends GenericException {
-
-    private static final String MESSAGE = "NOT_FOUND_EXCEPTION";
-    public NotFoundException(String detail) {
-        super(HttpStatus.NOT_FOUND.value(), MESSAGE, detail);
-    }
-}
