+++ src/main/java/ar/com/ddd/ddd_architecture/shared/dto/GenericResponse.java
@@ -1,0 +1,19 @@
+package ar.com.ddd.ddd_architecture.shared.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class GenericResponse<T> {
+    private Integer code;
+    private String message;
+    private T data;
+}
