+++ src/main/java/ar/com/ddd/ddd_architecture/shared/exceptions/handler/ExceptionHandlerController.java
@@ -1,0 +1,37 @@
+package ar.com.ddd.ddd_architecture.shared.exceptions.handler;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+import ar.com.ddd.ddd_architecture.shared.dto.ErrorResponseDto;
+import ar.com.ddd.ddd_architecture.shared.exceptions.InternalException;
+import ar.com.ddd.ddd_architecture.shared.exceptions.NotFoundException;
+
+
+@ControllerAdvice
+public class ExceptionHandlerController {
+    
+    @ExceptionHandler(InternalException.class)
+    public ResponseEntity<ErrorResponseDto> handleEmptyException(InternalException ex) {
+        return ResponseEntity
+        .status(ex.getCode())
+        .body(ErrorResponseDto.builder()
+                .code(ex.getCode())
+                .message(ex.getMessage())
+                .detail(ex.getDetail())
+                .build());
+    }
+
+    @ExceptionHandler(NotFoundException.class)
+    public ResponseEntity<ErrorResponseDto> handleNotFoundException(NotFoundException ex) {
+        return ResponseEntity
+        .status(ex.getCode())
+        .body(ErrorResponseDto.builder()
+                .code(ex.getCode())
+                .message(ex.getMessage())
+                .detail(ex.getDetail())
+                .build());
+    }
+
+}
