+++ src/main/java/ar/com/ddd/ddd_architecture/shared/utils/HttpWebClient.java
@@ -1,0 +1,14 @@
+package ar.com.ddd.ddd_architecture.shared.utils;
+import org.springframework.stereotype.Component;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+@Component
+public @interface HttpWebClient {
+    
+}
