--- src/main/resources/application.properties
@@ -1,6 +1,0 @@
-spring.application.name=ddd_architecture
-spring.datasource.username=root
-spring.datasource.password=
-spring.jpa.hibernate.ddl-auto=create-drop
-spring.datasource.url=jdbc:mysql://localhost:3307/ddd_database
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
